[IDLE-233] config: user-server RedisConfig, RedisClusterConfig 추가
